refactor(agent-runtime): move AgentCards from engine.spi to engine.a2a
AgentCards is an A2A card factory — not a core runtime SPI. Move it to the
a2a package alongside AgentCardProperties, AgentCardProvider, and
A2aClientAutoConfiguration.

engine.spi now contains only true core SPI types: AgentRuntimeHandler,
StreamAdapter, AgentExecutionResult, MemoryProvider, and Trajectory*.
