+++ agent-runtime/src/main/java/com/huawei/ascend/runtime/engine/versatile/VersatileAgentRuntimeHandler.java
@@ -2,7 +2,7 @@ package com.huawei.ascend.runtime.engine.versatile;
 
 import com.huawei.ascend.runtime.engine.AgentExecutionContext;
 import com.huawei.ascend.runtime.engine.a2a.AgentCardProvider;
-import com.huawei.ascend.runtime.engine.spi.AgentCards;
+import com.huawei.ascend.runtime.engine.a2a.AgentCards;
 import com.huawei.ascend.runtime.engine.spi.AgentRuntimeHandler;
 import com.huawei.ascend.runtime.engine.spi.StreamAdapter;
 import java.util.Map;
